% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_empty.r
\name{Data_empty}
\alias{Data_empty}
\title{Create a completely empty Data object}
\usage{
Data_empty(Data = NULL)
}
\arguments{
\item{Data}{Data object}
}
\description{
Create a completely empty Data object
}
\details{

}
\examples{
\dontrun{
Data_empty()
}
}
\author{
Nikolai Klibansky
}
\keyword{MSEtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
