% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_sub_dim.r
\name{Data_sub_dim}
\alias{Data_sub_dim}
\title{Subset MSEtool Data object from bam rdat object based on a particular dimension (i.e. year,)}
\usage{
Data_sub_dim(Data, nyr_rm = 0, type = "year")
}
\arguments{
\item{Data}{MSEtool Data object}

\item{nyr_rm}{number of years to remove from the end of the data series. used if type = "year".}

\item{type}{type of subsetting to do: "year" = trim year dimension. Right now, "year" is the only option, though "age" may be added in the future.}
}
\description{
Subset MSEtool Data object from bam rdat object based on a particular dimension (i.e. year,)
}
\examples{
\dontrun{
# Convert Black Sea Bass rdat to Data object
Data <- rdat_to_Data(rdat_BlackSeaBass)
Data_sub_dim(Data,1)
}
}
\author{
Nikolai Klibansky
}
\keyword{MSEtool}
\keyword{assessment}
\keyword{bam}
\keyword{fisheries}
\keyword{stock}
